causes/serializers.py
from rest_framework import serializers
from .models import Cause, Contribution

class CauseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Cause
        fields = '__all__'

class ContributionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contribution
        fields = ['name', 'email', 'amount']